requirements.txt
numpy
pandas
tensorflow
tqdm
scipy
dm-sonnet>=2.0.0b0
graph_nets
pre-commit
